requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
backcall==0.2.0
certifi==2020.6.20
cffi==1.14.4
chardet==4.0.0
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.2
decorator==4.4.2
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1
django-ckeditor==6.0.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-likes==2.0.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-secretballot==1.0.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.5.3
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
parso==0.7.1
pickleshare==0.7.5
Pillow==8.1.0
pipenv==2020.8.13
prompt-toolkit==3.0.8
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
Pygments==2.7.2
PyJWT==1.7.1
pylint==2.6.0
PyMySQL==1.0.2
pyserial==3.5
pytz==2020.1
PyYAML==5.3.1
requests==2.25.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
style==1.1.0
tabulate==0.8.7
toml==0.10.1
traitlets==5.0.5
update==0.0.1
uritemplate==3.0.1
urllib3==1.26.3
virtualenv==20.0.31
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.12.1
yoyo-migrations==7.2.1
